functional getPosition
