--- backend/src/websocket/getPosition/cubeLand/positionSetter.ts
@@ -1,4 +1,0 @@
-// Config
-const TIME = 120; // Seconds
-
-
